OCPBUGS-15910: After dual-stack conversion reconcile IPFamilies
This PR introduces a missing reconciliation of IPFamilies property in
the ControllerConfig. A consequence of the reconciliation missing is
that after conversion from single-stack to dual-stack the kubelet is
still configured as single-stack because the underlying MCO's variable
does not change its value.

With this PR, when converting such a cluster, ControllerConfig will
update the value and as a consequence reconfigure kubelet (specifically
`--node-ip` param).

Please note with this change we are introducing one reboot to the
process of cluster conversion.

Fixes: OCPBUGS-15910